new file level_map_scene.hpp + include fixes
